requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
astroid==2.5
async-timeout==3.0.1
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
discord.py==1.6.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pip-autoremove==0.9.1
pylint==2.7.0
python-dateutil==2.8.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
yarl==1.6.3
